Ambient-Health-PI/requirements.txt
python-kasa
fastapi
uvicorn[standard]
python-dotenv